requirements.txt
discord.py==2.3.2
PyYAML==6.0
PyNaCl==1.5.0
aiohttp==3.8.5
python-dotenv==1.0.0
